requirements.txt

autopep8==1.5.7
dj-database-url==0.5.0
pycodestyle==2.7.0
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0


Pillow==9.3.0
python-decouple==3.6
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
six==1.16.0
sqlparse==0.2.4

djongo==1.3.6

django-crispy-forms==1.14.0


celery~=5.2.7
requests~=2.28.1
shortuuid~=1.0.11